Move utility methods from HttpServletResponseExtension to MimeTypeUtils
